move to float and remove static declaration to global

+++ src/landscape.hpp
@@ -11,15 +11,15 @@ enum VegetationType {
   DRY
 } __attribute__((packed)); // packed so that it takes only 1 byte
 
-static_assert( sizeof(VegetationType) == 1 );
+static_assert(sizeof(VegetationType) == 1);
 
 struct Cell {
-  double elevation;
-  double wind_direction;
+  float elevation;
+  float wind_direction;
   bool burnable;
   VegetationType vegetation_type;
-  double fwi;
-  double aspect;
+  float fwi;
+  float aspect;
 };
 
 struct Landscape {
